fix(backend): correctifs review v0.2 setup (formule tilt-comp + race + nits)
- I1 : tilt_compensated_heading utilise la formule canonique AN-203
  (R = R_y(pitch) @ R_x(roll), m_world = R @ m_sensor) ; l'ancienne mélangeait
  les termes croisés et donnait ~0.87° d'erreur sur (pitch=5°, roll=10°,
  heading=0°). Test paramétré numpy ajouté pour couvrir 8 combos pitch+roll
  contre régression.
- I2 : _LazySensor.__aexit__ guarde adapter.stop() pour empêcher la dérive
  du refcount sur exception.
- I3 : adapter.start() reste sous le lock de session ; rollback de l'état
  si start() lève, pour éviter la course abort↔start.
- I4 : calibration_repo.get_offsets() catch ValidationError et dégrade en
  « non calibré » au lieu de propager 500.
- I5 : progress() snapshot atomique tous champs en locales avant de
  construire le modèle (évite vue incohérente pendant _clear_session).
- I7 : hz hors [1, 10] → 422 explicite (était silencieusement clampé).
- Nits : assert→RuntimeError dans les adapters (N2), commentaire O(N²)
  coverage_pct (N3), guard k < 1e-9 ellipsoide (N4), suppression
  except CancelledError redondant (N5), commentaire padding migrations
  (N7), try/finally finalize (N8), constante _TUBE_ZERO_ALT_DEG (N9),
  np.array(samples, dtype=float) bias_fit (N10).

+++ backend/astro_brain/services/_bias_fit.py
@@ -16,7 +16,7 @@ def compute_bias_and_sigma(
     if not samples:
         raise ValueError("samples must not be empty")
 
-    arr = np.array(samples)
+    arr = np.array(samples, dtype=float)
     mx, my, mz = arr.mean(axis=0).tolist()
     sigma = float(arr.std(axis=0).max())
     return (float(mx), float(my), float(mz)), sigma
